consolidate dataSourceCahps
