+++ services/ui-src/src/measures/2022/shared/CommonQuestions/index.tsx
@@ -2,7 +2,7 @@ export * from "shared/commonQuestions/CombinedRates";
 export * from "shared/commonQuestions/DefinitionsOfPopulation";
 export * from "shared/commonQuestions/DateRange";
 export * from "shared/commonQuestions/DataSource";
-export * from "./DataSourceCahps";
+export * from "shared/commonQuestions/DataSourceCahps";
 export * from "shared/commonQuestions/AdditionalNotes";
 export * from "./OtherPerformanceMeasure";
 export * from "./Reporting";
