--- services/ui-src/src/measures/2024/shared/CommonQuestions/DataSourceCahps/data.ts
@@ -1,26 +1,0 @@
-export interface OptionNode {
-  value: string;
-  subOptions?: {
-    label?: string;
-    options: OptionNode[];
-  };
-  description?: boolean;
-}
-
-export interface DataSourceData {
-  options: OptionNode[];
-  optionsLabel: string;
-}
-
-export const defaultData: DataSourceData = {
-  optionsLabel: "Which version of the CAHPS survey was used for reporting?",
-  options: [
-    {
-      value: "CAHPS 5.1H",
-    },
-    {
-      value: "Other Data Source",
-      description: true,
-    },
-  ],
-};
