+++ services/ui-src/src/measures/2024/shared/CommonQuestions/index.ts
@@ -2,7 +2,7 @@ export * from "shared/commonQuestions/CombinedRates";
 export * from "shared/commonQuestions/DefinitionsOfPopulation";
 export * from "shared/commonQuestions/DateRange";
 export * from "shared/commonQuestions/DataSource";
-export * from "./DataSourceCahps";
+export * from "shared/commonQuestions/DataSourceCahps";
 export * from "shared/commonQuestions/AdditionalNotes";
 export * from "./OtherPerformanceMeasure";
 export * from "./Reporting";
