+++ services/ui-src/src/shared/commonQuestions/DataSourceCahps.tsx
@@ -1,8 +1,11 @@
 import * as QMR from "components";
 import { useCustomRegister } from "hooks/useCustomRegister";
 import * as Types from "../types";
-import { DataSourceData, defaultData } from "./data";
+import { useContext } from "react";
+import { DataSourceData, defaultData } from "../types/TypeDataSourceCahps";
+import { parseLabelToHTML } from "utils/parser";
 import * as DC from "dataConstants";
+import SharedContext from "shared/SharedContext";
 
 interface DataSourceProps {
   data?: DataSourceData;
@@ -13,6 +16,8 @@ interface DataSourceProps {
  */
 export const DataSourceRadio = ({ data = defaultData }: DataSourceProps) => {
   const register = useCustomRegister<Types.DataSource>();
+
+  const labels: any = useContext(SharedContext);
 
   return (
     <QMR.CoreQuestionWrapper testid="data-source" label="Data Source">
@@ -30,16 +35,7 @@ export const DataSourceRadio = ({ data = defaultData }: DataSourceProps) => {
             children: [
               <QMR.TextArea
                 name="describeTheDataSource"
-                label={
-                  <>
-                    Describe the data source (
-                    <em>
-                      text in this field is included in publicly-reported
-                      state-specific comments
-                    </em>
-                    ):
-                  </>
-                }
+                label={parseLabelToHTML(labels?.DataSource.describeDataSrc)}
                 key="dataSourceOtherTextArea"
                 formLabelProps={{
                   fontWeight: "normal",
